refactor(auth): extract authorize/device pages, add Disclosure primitive
- extract /authorize and /device UI from route files into components/auth/
  (routes are now thin shells: route config + component import)
- new components/ui/chevron-icon.tsx — single canonical chevron used
  everywhere disclosure or dropdown triggers appear
- new components/ui/disclosure.tsx — shared collapsible row primitive
  (fixes square-hover + white-hover bugs at the root for all uses)
- account-permissions-input Models row → Disclosure
- authorize Permissions summary + period-picker pill → ChevronIcon
- auth modal: amber backdrop instead of green dim (data-theme=amber on
  outer wrapper so the whole screen reads amber)
- permission-ui input: add focus-visible:outline-none + ring-2 width so
  the amber focus ring actually renders (was eclipsed by browser blue)
- design-showcase: Disclosure section added
